remove ammonite dep
